Switched back to POST requests
GET is fine for most requests, but things like uploading playlists or updating the Jukebox play queue can include thousands of song IDs which while technically in spec, my not be handled by Subsonic API servers or things like Nginx reverse proxies, etc as most HTTP servers tend to have reasonable limits on url + query length (like 500-2000 bytes or so). Since POST works fine, it's simpler to just always use it rather than only using it when the query string is too long.

If it ends up causing issues with non-Subsonic servers (aka anything that implements the API and isn't Subsonic or a fork), I'll switch it to use GET by default and only use POST when larger than 500 bytes or so.